feat: add UT for controller and service for mrf
