+++ src/app/modules/submission/multirespondent-submission/__tests__/multirespondent-submission.controller.spec.ts
@@ -1,141 +1,71 @@
-import dbHandler from '__tests__/unit/backend/helpers/jest-db'
 import expressHandler from '__tests__/unit/backend/helpers/jest-express'
 import { ObjectId } from 'bson'
-import { merge } from 'lodash'
-import mongoose from 'mongoose'
-import { ok, okAsync } from 'neverthrow'
-import {
-  BasicField,
-  FormAuthType,
-  FormWorkflowStepDto,
-  WorkflowType,
-} from 'shared/types'
+import { merge, omit } from 'lodash'
+import { Types } from 'mongoose'
+import { errAsync, ok, okAsync } from 'neverthrow'
+import { FormAuthType } from 'shared/types'
 
-import { getMultirespondentSubmissionModel } from 'src/app/models/submission.server.model'
 import * as FormService from 'src/app/modules/form/form.service'
-import MailService from 'src/app/services/mail/mail.service'
+import { MailSendError } from 'src/app/services/mail/mail.errors'
+import { IMultirespondentSubmissionSchema } from 'src/types'
 
+import {
+  AttachmentUploadError,
+  InvalidWorkflowTypeError,
+  SubmissionNotFoundError,
+  SubmissionSaveError,
+} from '../../submission.errors'
 import {
   submitMultirespondentFormForTest,
   updateMultirespondentSubmissionForTest,
 } from '../multirespondent-submission.controller'
+import * as MultiRespondentSubmissionService from '../multirespondent-submission.service'
 
 jest.mock('src/app/modules/datadog/datadog.utils')
 
-const MultiRespondentSubmission = getMultirespondentSubmissionModel(mongoose)
-
+jest.mock('src/app/modules/form/form.service')
 const MockFormService = jest.mocked(FormService)
+jest.mock(
+  'src/app/modules/submission/multirespondent-submission/multirespondent-submission.service',
+)
+const MockMultiRespondentSubmissionService = jest.mocked(
+  MultiRespondentSubmissionService,
+)
 
-describe('multirespondent-submission.controller', () => {
-  beforeAll(async () => {
-    await dbHandler.connect()
+const mockFormId = new ObjectId().toHexString()
+const mockMrfForm = {
+  _id: mockFormId,
+  workflow: [],
+}
+const mockSubmissionId = new ObjectId().toHexString()
+const mockMrfSubmission = {
+  _id: mockSubmissionId,
+} as IMultirespondentSubmissionSchema & { _id: Types.ObjectId }
 
+describe('multiresponodent-submision.controller', () => {
+  beforeEach(() => {
     MockFormService.isFormPublic = jest.fn().mockReturnValue(ok(true))
-    MockFormService.checkIsIntranetFormAccess = jest.fn().mockReturnValue(false)
     MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
       .fn()
       .mockReturnValue(okAsync(mockMrfForm))
+
+    MockMultiRespondentSubmissionService.createMultiRespondentFormSubmission =
+      jest.fn().mockReturnValue(okAsync(mockMrfSubmission))
+    MockMultiRespondentSubmissionService.updateMultiRespondentFormSubmission =
+      jest.fn().mockReturnValue(okAsync(mockMrfSubmission))
+    MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionCreateActions =
+      jest.fn().mockReturnValue(okAsync(true))
+    MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionUpdateActions =
+      jest.fn().mockReturnValue(okAsync(true))
   })
 
-  afterEach(async () => {
+  afterEach(() => {
     jest.clearAllMocks()
-    await dbHandler.clearDatabase()
   })
 
-  afterAll(async () => {
-    await dbHandler.closeDatabase()
-  })
-
-  const mockFormId = new ObjectId().toHexString()
-  const mockMrfForm = {
-    _id: mockFormId,
-    workflow: [],
-  }
-
-  const mockSubmissionId = new ObjectId().toHexString()
-
-  describe('mrf approval email notification when approval step exists', () => {
-    it('workflow continues and does not send approved outcome email when mrf is approved for mid step of multiple step MRF', async () => {
+  describe('submitMultirespondentForm', () => {
+    it('returns 200 ok when form validation passes and invokes createMultiRespondentFormSubmission and performMultiRespondentPostSubmissionCreateActions', async () => {
       // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-      const sendMrfApprovalEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfApprovalEmail',
-      )
-      const sendMRFWorkflowStepEmailSpy = jest.spyOn(
-        MailService,
-        'sendMRFWorkflowStepEmail',
-      )
-
-      const expectedEmails = ['expected1@example.com']
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const yesNoFieldId1 = new ObjectId().toHexString()
-      const yesNoFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: 'not_expected_1@example.com',
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: expectedEmails[0],
-          },
-        },
-        [yesNoFieldId1]: {
-          fieldType: BasicField.YesNo,
-          answer: 'Yes',
-        },
-        [yesNoFieldId2]: {
-          fieldType: BasicField.YesNo,
-          answer: 'No',
-        },
-      }
-
-      const threeStepApprovalWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_expected_2@example.com'],
-          edit: [yesNoFieldId1],
-          approval_field: yesNoFieldId1,
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [yesNoFieldId2],
-          // no approval field for last step
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
-
       const mockReq = expressHandler.mockRequest({
         params: {
           formId: mockFormId,
@@ -149,671 +79,62 @@ describe('multirespondent-submission.controller', () => {
             _id: mockFormId,
             authType: FormAuthType.NIL,
             getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: threeStepApprovalWorkflow,
-            emails: [expectedEmails[1]],
-            stepsToNotify: [stepOneId],
           },
           encryptedPayload: {
             encryptedContent: 'encryptedContent',
             version: 1,
             submissionPublicKey: 'submissionPublicKey',
             encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: 1, // 2nd step of 3 steps workflow
+            responses: {},
+            workflowStep: 0,
           },
         } as any,
       })
-      const mockRes = expressHandler.mockResponse({})
-
-      // Act
-      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
-      // Assert
-      // next workflow step email is sent only
-      expect(sendMrfApprovalEmailSpy).not.toHaveBeenCalled()
-      expect(sendMrfWorkflowCompletionEmailSpy).not.toHaveBeenCalled()
-      expect(sendMRFWorkflowStepEmailSpy).toHaveBeenCalledTimes(1)
-      // destination emails are correct
-      expect(
-        sendMRFWorkflowStepEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(expectedEmails)
-      expect(sendMRFWorkflowStepEmailSpy.mock.calls[0][0].emails.length).toBe(
-        expectedEmails.length,
-      )
-    })
-
-    it('sends approved outcome email when mrf has approval step earlier but last step is not approval step', async () => {
-      // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-      const sendMrfApprovalEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfApprovalEmail',
-      )
-      const sendMRFWorkflowStepEmailSpy = jest.spyOn(
-        MailService,
-        'sendMRFWorkflowStepEmail',
-      )
-
-      const expectedEmails = ['expected1@example.com', 'expected2@example.com']
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const yesNoFieldId1 = new ObjectId().toHexString()
-      const yesNoFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: expectedEmails[0],
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: 'not_expected_1@example.com',
-          },
-        },
-        [yesNoFieldId1]: {
-          fieldType: BasicField.YesNo,
-          answer: 'Yes',
-        },
-        [yesNoFieldId2]: {
-          fieldType: BasicField.YesNo,
-          answer: 'No',
-        },
-      }
-
-      const threeStepApprovalWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_expected_2@example.com'],
-          edit: [yesNoFieldId1],
-          approval_field: yesNoFieldId1,
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [yesNoFieldId2],
-          // no approval field for last step
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
-
-      const mockReq = expressHandler.mockRequest({
-        params: {
-          formId: mockFormId,
-          submissionId: mockSubmissionId,
-        },
-        body: {} as any,
-      })
-      const mockSubmitMrfReq = merge(mockReq, {
-        formsg: {
-          formDef: {
-            _id: mockFormId,
-            authType: FormAuthType.NIL,
-            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: threeStepApprovalWorkflow,
-            emails: [expectedEmails[1]],
-            stepsToNotify: [stepOneId],
-          },
-          encryptedPayload: {
-            encryptedContent: 'encryptedContent',
-            version: 1,
-            submissionPublicKey: 'submissionPublicKey',
-            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: threeStepApprovalWorkflow.length - 1, // last step
-          },
-        } as any,
-      })
-      const mockRes = expressHandler.mockResponse({})
-
-      // Act
-      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
-      // Assert
-      // approval email is sent instead of completion email
-      expect(sendMrfApprovalEmailSpy).toHaveBeenCalledTimes(1)
-      expect(sendMrfWorkflowCompletionEmailSpy).not.toHaveBeenCalled()
-      expect(sendMRFWorkflowStepEmailSpy).not.toHaveBeenCalled()
-      // is approve email and destination emails are correct
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].isRejected).toBeFalse()
-      expect(
-        sendMrfApprovalEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(expectedEmails)
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].emails.length).toBe(
-        expectedEmails.length,
-      )
-    })
-
-    it('sends approved outcome email when mrf is approved for last step of multiple step MRF', async () => {
-      // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-      const sendMrfApprovalEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfApprovalEmail',
-      )
-      const sendMRFWorkflowStepEmailSpy = jest.spyOn(
-        MailService,
-        'sendMRFWorkflowStepEmail',
-      )
-
-      const expectedEmails = ['expected1@example.com', 'expected2@example.com']
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const yesNoFieldId1 = new ObjectId().toHexString()
-      const yesNoFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: expectedEmails[0],
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: 'not_expected_1@example.com',
-          },
-        },
-        [yesNoFieldId1]: {
-          fieldType: BasicField.YesNo,
-          answer: 'Yes',
-        },
-        [yesNoFieldId2]: {
-          fieldType: BasicField.YesNo,
-          answer: 'Yes',
-        },
-      }
-
-      const threeStepApprovalWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_expected_2@example.com'],
-          edit: [yesNoFieldId1],
-          approval_field: yesNoFieldId1,
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [yesNoFieldId2],
-          approval_field: yesNoFieldId2,
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
-
-      const mockReq = expressHandler.mockRequest({
-        params: {
-          formId: mockFormId,
-          submissionId: mockSubmissionId,
-        },
-        body: {} as any,
-      })
-      const mockSubmitMrfReq = merge(mockReq, {
-        formsg: {
-          formDef: {
-            _id: mockFormId,
-            authType: FormAuthType.NIL,
-            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: threeStepApprovalWorkflow,
-            emails: [expectedEmails[1]],
-            stepsToNotify: [stepOneId],
-          },
-          encryptedPayload: {
-            encryptedContent: 'encryptedContent',
-            version: 1,
-            submissionPublicKey: 'submissionPublicKey',
-            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: threeStepApprovalWorkflow.length - 1, // last step
-          },
-        } as any,
-      })
-      const mockRes = expressHandler.mockResponse({})
-
-      // Act
-      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
-      // Assert
-      // approval email is sent instead of completion email
-      expect(sendMrfApprovalEmailSpy).toHaveBeenCalledTimes(1)
-      expect(sendMrfWorkflowCompletionEmailSpy).not.toHaveBeenCalled()
-      expect(sendMRFWorkflowStepEmailSpy).not.toHaveBeenCalled()
-      // is approve email and destination emails are correct
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].isRejected).toBeFalse()
-      expect(
-        sendMrfApprovalEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(expectedEmails)
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].emails.length).toBe(
-        expectedEmails.length,
-      )
-    })
-
-    it('workflow terminates and sends not approved outcome email when mrf is rejected for mid step of multiple step MRF', async () => {
-      // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-      const sendMrfApprovalEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfApprovalEmail',
-      )
-      const sendMRFWorkflowStepEmailSpy = jest.spyOn(
-        MailService,
-        'sendMRFWorkflowStepEmail',
-      )
-
-      const expectedEmails = ['expected1@example.com', 'expected2@example.com']
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const yesNoFieldId1 = new ObjectId().toHexString()
-      const yesNoFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: expectedEmails[0],
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: 'not_expected_1@example.com',
-          },
-        },
-        [yesNoFieldId1]: {
-          fieldType: BasicField.YesNo,
-          answer: 'No',
-        },
-      }
-
-      const threeStepApprovalWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_expected_2@example.com'],
-          edit: [yesNoFieldId1],
-          approval_field: yesNoFieldId1,
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [yesNoFieldId2],
-          approval_field: yesNoFieldId2,
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
-
-      const mockReq = expressHandler.mockRequest({
-        params: {
-          formId: mockFormId,
-          submissionId: mockSubmissionId,
-        },
-        body: {} as any,
-      })
-      const mockSubmitMrfReq = merge(mockReq, {
-        formsg: {
-          formDef: {
-            _id: mockFormId,
-            authType: FormAuthType.NIL,
-            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: threeStepApprovalWorkflow,
-            emails: [expectedEmails[1]],
-            stepsToNotify: [stepOneId],
-          },
-          encryptedPayload: {
-            encryptedContent: 'encryptedContent',
-            version: 1,
-            submissionPublicKey: 'submissionPublicKey',
-            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: 1, // 2nd step of 3 steps workflow
-          },
-        } as any,
-      })
-      const mockRes = expressHandler.mockResponse({})
-
-      // Act
-      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
-      // Assert
-      // approval email is sent instead of completion email
-      expect(sendMrfApprovalEmailSpy).toHaveBeenCalledTimes(1)
-      expect(sendMrfWorkflowCompletionEmailSpy).not.toHaveBeenCalled()
-      expect(sendMRFWorkflowStepEmailSpy).not.toHaveBeenCalled()
-      // is rejected email and destination emails are correct
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].isRejected).toBeTrue()
-      expect(
-        sendMrfApprovalEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(expectedEmails)
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].emails.length).toBe(
-        expectedEmails.length,
-      )
-    })
-
-    it('workflow terminates and sends not approved outcome email when mrf is rejected for last step of multiple step MRF', async () => {
-      // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-      const sendMrfApprovalEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfApprovalEmail',
-      )
-      const sendMRFWorkflowStepEmailSpy = jest.spyOn(
-        MailService,
-        'sendMRFWorkflowStepEmail',
-      )
-
-      const expectedEmails = ['expected1@example.com', 'expected2@example.com']
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const yesNoFieldId1 = new ObjectId().toHexString()
-      const yesNoFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: expectedEmails[0],
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: BasicField.Email,
-          answer: {
-            value: 'not_expected_1@example.com',
-          },
-        },
-        [yesNoFieldId1]: {
-          fieldType: BasicField.YesNo,
-          answer: 'Yes',
-        },
-        [yesNoFieldId2]: {
-          fieldType: BasicField.YesNo,
-          answer: 'No',
-        },
-      }
-
-      const threeStepApprovalWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_expected_2@example.com'],
-          edit: [yesNoFieldId1],
-          approval_field: yesNoFieldId1,
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [yesNoFieldId2],
-          approval_field: yesNoFieldId2,
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
-
-      const mockReq = expressHandler.mockRequest({
-        params: {
-          formId: mockFormId,
-          submissionId: mockSubmissionId,
-        },
-        body: {} as any,
-      })
-      const mockSubmitMrfReq = merge(mockReq, {
-        formsg: {
-          formDef: {
-            _id: mockFormId,
-            authType: FormAuthType.NIL,
-            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: threeStepApprovalWorkflow,
-            emails: [expectedEmails[1]],
-            stepsToNotify: [stepOneId],
-          },
-          encryptedPayload: {
-            encryptedContent: 'encryptedContent',
-            version: 1,
-            submissionPublicKey: 'submissionPublicKey',
-            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: threeStepApprovalWorkflow.length - 1, // last step
-          },
-        } as any,
-      })
-      const mockRes = expressHandler.mockResponse({})
-
-      // Act
-      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
-      // Assert
-      // approval email is sent instead of completion email
-      expect(sendMrfApprovalEmailSpy).toHaveBeenCalledTimes(1)
-      expect(sendMrfWorkflowCompletionEmailSpy).not.toHaveBeenCalled()
-      expect(sendMRFWorkflowStepEmailSpy).not.toHaveBeenCalled()
-      // is rejected email and destination emails are correct
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].isRejected).toBeTrue()
-      expect(
-        sendMrfApprovalEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(expectedEmails)
-      expect(sendMrfApprovalEmailSpy.mock.calls[0][0].emails.length).toBe(
-        expectedEmails.length,
-      )
-    })
-  })
-
-  describe('mrf completion email notification when no approval step exists', () => {
-    it('sends completion email when single step mrf is completed', async () => {
-      // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-
-      const singleStepWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: new ObjectId().toHexString(),
-          workflow_type: WorkflowType.Static,
-          emails: [],
-          edit: [],
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      const mockReq = expressHandler.mockRequest({
-        params: { formId: mockFormId },
-        body: {} as any,
-      })
-      const mockSubmitMrfReq = merge(mockReq, {
-        formsg: {
-          formDef: {
-            _id: mockFormId,
-            authType: FormAuthType.NIL,
-            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: singleStepWorkflow,
-            emails: ['email1@example.com'],
-          },
-          encryptedPayload: {
-            encryptedContent: 'encryptedContent',
-            version: 1,
-            submissionPublicKey: 'submissionPublicKey',
-            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-          },
-        } as any,
-      })
-      const mockRes = expressHandler.mockResponse({})
+      const mockRes = expressHandler.mockResponse()
 
       // Act
       await submitMultirespondentFormForTest(mockSubmitMrfReq, mockRes)
 
       // Assert
-      expect(sendMrfWorkflowCompletionEmailSpy).toHaveBeenCalledTimes(1)
+      // save submission is invoked with correct args
       expect(
-        sendMrfWorkflowCompletionEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(['email1@example.com'])
+        MockMultiRespondentSubmissionService.createMultiRespondentFormSubmission,
+      ).toHaveBeenCalledOnce()
       expect(
-        sendMrfWorkflowCompletionEmailSpy.mock.calls[0][0].emails.length,
-      ).toBe(1)
+        omit(
+          MockMultiRespondentSubmissionService
+            .createMultiRespondentFormSubmission.mock.calls[0][0],
+          'logMeta',
+        ),
+      ).toEqual({
+        encryptedPayload: mockSubmitMrfReq.formsg.encryptedPayload,
+        form: mockSubmitMrfReq.formsg.formDef,
+      })
+
+      // Assert post save actions are invoked with correct args
+      expect(
+        MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionCreateActions,
+      ).toHaveBeenCalledOnce()
+      expect(
+        omit(
+          MockMultiRespondentSubmissionService
+            .performMultiRespondentPostSubmissionCreateActions.mock.calls[0][0],
+          'logMeta',
+        ),
+      ).toEqual({
+        form: mockSubmitMrfReq.formsg.formDef,
+        encryptedPayload: mockSubmitMrfReq.formsg.encryptedPayload,
+        submissionId: mockSubmissionId,
+      })
+      // Expect 200 ok
+      expect(mockRes.status).not.toHaveBeenCalled() // default is 200 ok
     })
 
-    it('sends completion email when multi-step mrf is completed and only to specified steps only and also static emails', async () => {
+    it('returns 400 bad request if attachment upload error occurs when createMultiRespondentFormSubmission', async () => {
       // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-
-      const expectedEmails = [
-        'expected1@example.com',
-        'expected2@example.com',
-        'expected3@example.com',
-        'expected4@example.com',
-      ]
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-      const stepFourId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: 'email',
-          answer: {
-            value: expectedEmails[0],
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: 'email',
-          answer: {
-            value: 'not_expected_1@example.com',
-          },
-        },
-      }
-
-      const fourStepWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_expected_2@example.com'],
-          edit: [],
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [],
-        },
-        {
-          _id: stepFourId,
-          workflow_type: WorkflowType.Static,
-          emails: [expectedEmails[1], expectedEmails[2]],
-          edit: [],
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
+      const attachmentUploadError = new AttachmentUploadError()
+      MockMultiRespondentSubmissionService.createMultiRespondentFormSubmission =
+        jest.fn().mockReturnValue(errAsync(attachmentUploadError))
 
       const mockReq = expressHandler.mockRequest({
         params: {
@@ -828,107 +149,35 @@ describe('multirespondent-submission.controller', () => {
             _id: mockFormId,
             authType: FormAuthType.NIL,
             getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: fourStepWorkflow,
-            emails: [expectedEmails[3]],
-            stepsToNotify: [stepOneId, stepFourId],
           },
           encryptedPayload: {
             encryptedContent: 'encryptedContent',
             version: 1,
             submissionPublicKey: 'submissionPublicKey',
             encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: fourStepWorkflow.length - 1, // last step
+            responses: {},
+            workflowStep: 0,
           },
         } as any,
       })
-      const mockRes = expressHandler.mockResponse({})
+      const mockRes = expressHandler.mockResponse()
 
       // Act
-      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
+      await submitMultirespondentFormForTest(mockSubmitMrfReq, mockRes)
 
       // Assert
-      expect(sendMrfWorkflowCompletionEmailSpy).toHaveBeenCalledTimes(1)
-      // The emails sent to should only be the expected emails exactly
-      expect(
-        sendMrfWorkflowCompletionEmailSpy.mock.calls[0][0].emails,
-      ).toContainAllValues(expectedEmails)
-      expect(
-        sendMrfWorkflowCompletionEmailSpy.mock.calls[0][0].emails.length,
-      ).toBe(expectedEmails.length)
+      expect(mockRes.status).toHaveBeenCalledWith(400)
+      expect(mockRes.json).toHaveBeenCalledWith({
+        message:
+          'Could not upload attachments for submission. For assistance, please contact the person who asked you to fill in this form.',
+      })
     })
 
-    it('does not send completion email when step number >0 and mrf not completed', async () => {
+    it('returns 500 internal server error when submission fails to save', async () => {
       // Arrange
-      const sendMrfWorkflowCompletionEmailSpy = jest.spyOn(
-        MailService,
-        'sendMrfWorkflowCompletionEmail',
-      )
-
-      const selectedEmails = [
-        'seelcted1@example.com',
-        'seelcted2@example.com',
-        'seelcted3@example.com',
-        'seelcted4@example.com',
-      ]
-
-      const stepOneId = new ObjectId().toHexString()
-      const stepTwoId = new ObjectId().toHexString()
-      const stepThreeId = new ObjectId().toHexString()
-      const stepFourId = new ObjectId().toHexString()
-
-      const emailFieldId1 = new ObjectId().toHexString()
-      const emailFieldId2 = new ObjectId().toHexString()
-
-      const submissionResponses = {
-        [emailFieldId1]: {
-          fieldType: 'email',
-          answer: {
-            value: selectedEmails[0],
-          },
-        },
-        [emailFieldId2]: {
-          fieldType: 'email',
-          answer: {
-            value: 'not_selected_1@example.com',
-          },
-        },
-      }
-
-      const fourStepWorkflow: FormWorkflowStepDto[] = [
-        {
-          _id: stepOneId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId1,
-          edit: [],
-        },
-        {
-          _id: stepTwoId,
-          workflow_type: WorkflowType.Static,
-          emails: ['not_selected_2@example.com'],
-          edit: [],
-        },
-        {
-          _id: stepThreeId,
-          workflow_type: WorkflowType.Dynamic,
-          field: emailFieldId2,
-          edit: [],
-        },
-        {
-          _id: stepFourId,
-          workflow_type: WorkflowType.Static,
-          emails: [selectedEmails[1], selectedEmails[2]],
-          edit: [],
-        },
-      ]
-
-      MockFormService.checkFormSubmissionLimitAndDeactivateForm = jest
-        .fn()
-        .mockReturnValue(okAsync(mockMrfForm))
-
-      MultiRespondentSubmission.findById = jest.fn().mockReturnValue({
-        save: () => true,
-      })
+      const submissionSaveError = new SubmissionSaveError()
+      MockMultiRespondentSubmissionService.createMultiRespondentFormSubmission =
+        jest.fn().mockReturnValue(errAsync(submissionSaveError))
 
       const mockReq = expressHandler.mockRequest({
         params: {
@@ -943,27 +192,378 @@ describe('multirespondent-submission.controller', () => {
             _id: mockFormId,
             authType: FormAuthType.NIL,
             getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
-            workflow: fourStepWorkflow,
-            emails: [selectedEmails[3]],
-            stepsToNotify: [stepOneId, stepFourId],
           },
           encryptedPayload: {
             encryptedContent: 'encryptedContent',
             version: 1,
             submissionPublicKey: 'submissionPublicKey',
             encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
-            responses: submissionResponses,
-            workflowStep: fourStepWorkflow.length - 2, // not last step
+            responses: {},
+            workflowStep: 0,
           },
         } as any,
       })
-      const mockRes = expressHandler.mockResponse({})
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await submitMultirespondentFormForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).toHaveBeenCalledWith(500)
+      expect(mockRes.json).toHaveBeenCalledWith({
+        message: submissionSaveError.message,
+      })
+    })
+
+    it('returns 200 ok when step has invalid workflow type', async () => {
+      // Arrange
+      const invalidWorkflowTypeError = new InvalidWorkflowTypeError()
+      MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionCreateActions =
+        jest.fn().mockReturnValue(errAsync(invalidWorkflowTypeError))
+
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 0,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await submitMultirespondentFormForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).not.toHaveBeenCalled() // default is 200 ok
+    })
+
+    it('returns 200 ok when mail send error occurs', async () => {
+      // Arrange
+      const mailSendError = new MailSendError()
+      MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionCreateActions =
+        jest.fn().mockReturnValue(errAsync(mailSendError))
+
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 0,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await submitMultirespondentFormForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).not.toHaveBeenCalled() // default is 200 ok
+    })
+  })
+
+  describe('updateMultirespondentSubmission', () => {
+    it('returns 200 ok when form validation passes and invokes updateMultiRespondentFormSubmission and performMultiRespondentPostSubmissionUpdateActions', async () => {
+      // Arrange
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 0,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
 
       // Act
       await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
 
       // Assert
-      expect(sendMrfWorkflowCompletionEmailSpy).not.toHaveBeenCalled()
+      // save submission is invoked with correct args
+      expect(
+        MockMultiRespondentSubmissionService.updateMultiRespondentFormSubmission,
+      ).toHaveBeenCalledOnce()
+      expect(
+        omit(
+          MockMultiRespondentSubmissionService
+            .updateMultiRespondentFormSubmission.mock.calls[0][0],
+          'logMeta',
+        ),
+      ).toEqual({
+        formId: mockFormId,
+        submissionId: mockSubmissionId,
+        encryptedPayload: mockSubmitMrfReq.formsg.encryptedPayload,
+        form: mockSubmitMrfReq.formsg.formDef,
+      })
+
+      // Assert post save actions are invoked with correct args
+      expect(
+        MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionUpdateActions,
+      ).toHaveBeenCalledOnce()
+      expect(
+        omit(
+          MockMultiRespondentSubmissionService
+            .performMultiRespondentPostSubmissionUpdateActions.mock.calls[0][0],
+          'logMeta',
+        ),
+      ).toEqual({
+        form: mockSubmitMrfReq.formsg.formDef,
+        encryptedPayload: mockSubmitMrfReq.formsg.encryptedPayload,
+        submissionId: mockSubmissionId,
+      })
+      // Expect 200 ok
+      expect(mockRes.status).not.toHaveBeenCalled() // default is 200 ok
+    })
+
+    it('returns 400 bad request when attachment upload fails when updateMultiRespondentFormSubmission', async () => {
+      // Arrange
+      const attachmentUploadError = new AttachmentUploadError()
+      MockMultiRespondentSubmissionService.updateMultiRespondentFormSubmission =
+        jest.fn().mockReturnValue(errAsync(attachmentUploadError))
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 1,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).toHaveBeenCalledWith(400)
+      expect(mockRes.json).toHaveBeenCalledWith({
+        message:
+          'Could not upload attachments for submission. For assistance, please contact the person who asked you to fill in this form.',
+      })
+    })
+
+    it('returns 500 internal server error when submission fails to save', async () => {
+      // Arrange
+      const submissionSaveError = new SubmissionSaveError()
+      MockMultiRespondentSubmissionService.updateMultiRespondentFormSubmission =
+        jest.fn().mockReturnValue(errAsync(submissionSaveError))
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 1,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).toHaveBeenCalledWith(500)
+      expect(mockRes.json).toHaveBeenCalledWith({
+        message: submissionSaveError.message,
+        submissionId: mockSubmissionId,
+      })
+    })
+
+    it('returns 404 not found when submission id not found', async () => {
+      // Arrange
+      const submissionNotFoundError = new SubmissionNotFoundError()
+      MockMultiRespondentSubmissionService.updateMultiRespondentFormSubmission =
+        jest.fn().mockReturnValue(errAsync(submissionNotFoundError))
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 1,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).toHaveBeenCalledWith(404)
+      expect(mockRes.json).toHaveBeenCalledWith({
+        message: submissionNotFoundError.message,
+      })
+    })
+
+    it('returns 200 ok when mail send error occurs', async () => {
+      // Arrange
+      const mailSendError = new MailSendError()
+      MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionUpdateActions =
+        jest.fn().mockReturnValue(errAsync(mailSendError))
+
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 1,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).not.toHaveBeenCalled() // default is 200 ok
+    })
+
+    it('returns 200 ok when step has invalid workflow type', async () => {
+      // Arrange
+      const invalidWorkflowTypeError = new InvalidWorkflowTypeError()
+      MockMultiRespondentSubmissionService.performMultiRespondentPostSubmissionUpdateActions =
+        jest.fn().mockReturnValue(errAsync(invalidWorkflowTypeError))
+
+      const mockReq = expressHandler.mockRequest({
+        params: {
+          formId: mockFormId,
+          submissionId: mockSubmissionId,
+        },
+        body: {} as any,
+      })
+      const mockSubmitMrfReq = merge(mockReq, {
+        formsg: {
+          formDef: {
+            _id: mockFormId,
+            authType: FormAuthType.NIL,
+            getUniqueMyInfoAttrs: jest.fn().mockReturnValue([]),
+          },
+          encryptedPayload: {
+            encryptedContent: 'encryptedContent',
+            version: 1,
+            submissionPublicKey: 'submissionPublicKey',
+            encryptedSubmissionSecretKey: 'encryptedSubmissionSecretKey',
+            responses: {},
+            workflowStep: 1,
+          },
+        } as any,
+      })
+      const mockRes = expressHandler.mockResponse()
+
+      // Act
+      await updateMultirespondentSubmissionForTest(mockSubmitMrfReq, mockRes)
+
+      // Assert
+      expect(mockRes.status).not.toHaveBeenCalled() // default is 200 ok
     })
   })
 })
